# back-end/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorLogResponse:
    description: Standard error response body containing details about the error occurrence
    properties:
      call:
        description: The API endpoint/URL that was called
        example: /getpoint/123
        type: string
      debuginfo:
        description: Detailed debug information about the error
        example: Point with ID 123 not found
        type: string
      error:
        description: Short error title/description
        example: Not Found
        type: string
      originator:
        description: IP address or identifier of the request originator
        example: 192.168.1.100
        type: string
      return_code:
        description: HTTP status code returned
        example: 404
        type: integer
      timeref:
        description: 'Timestamp when the error occurred (format: YYYY-MM-DD HH:MM)'
        example: 2025-12-25 14:30
        type: string
    type: object
...
  handlers.UpdatePointRequest:
    description: Request body for updating charging point properties
    properties:
      is_manual_price:
        description: Whether the price is manually set or dynamically calculated
        example: true
        type: boolean
      kwhprice:
        description: Price per kWh in the local currency
        example: 0.35
        type: number
      status:
        description: Status of the charging point (available, charging, reserved,
          malfunction, offline)
        example: available
        type: string
    type: object
  handlers.UpdatePointResponse:
    description: Response body containing the updated charging point details
    properties:
      is_manual_price:
        description: Indicates if the price is manually set
        example: true
        type: boolean
      kwhprice:
        description: Current price per kWh
        example: 0.35
        type: number
      pointid:
        description: Unique identifier of the charging point
        example: "123"
        type: string
      status:
        description: Current status of the charging point
        example: available
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing electric vehicle charging points
  title: EV Charging API
  version: "1.0"
paths:
...
  /updpoint/{id}:
    post:
      consumes:
      - application/json
      description: |-
        Updates the properties of a specific charging point including status, price, and pricing mode.
        At least one field (status, kwhprice, or is_manual_price) must be provided.
        When setting kwhprice without specifying is_manual_price, it defaults to manual pricing mode.
      parameters:
      - description: Charging Point ID
        in: path
        name: id
        required: true
        type: string
      - description: Update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated charging point
          schema:
            $ref: '#/definitions/handlers.UpdatePointResponse'
        "400":
          description: Bad Request - Invalid input or missing required fields
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "404":
          description: Not Found - Charging point does not exist
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "500":
          description: Internal Server Error - Database error
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
      summary: Update a charging point
      tags:
      - Points
swagger: "2.0"